+++ CHANGELOG.md
@@ -5,6 +5,24 @@
 # Changelog
 All notable changes to this project will be documented here.
 
+## Version [0.9.8.0] -
+
+[Added] ```getUserBots``` method
+
+[Added] ```onUserBots``` callback
+
+[Added] ```blockAssistant``` method
+
+[Added] ```onAssistantBlocked``` callback
+
+[Added] ```unBlockAssistant``` method
+
+[Added] ```onAssistantUnBlocked``` callback
+
+[Added] ```getBlocksAssistant``` method
+
+[Added] ```onAssistantBlocks``` callback
+
 ## Version [0.9.7.0] -2020-2-24
 
 [Added] ```getAssistantHistory``` method